requirements.txt
Flask==3.0.0
gunicorn==21.2.0
grpcio==1.73.1
grpcio-tools==1.73.1
protobuf==6.31.1